Feat: Use react-query to get my email address
